Add Spree::Order#create_shipments_for_line_item
Sometimes necessity arises to create shipments for an order that is already
completed, for example when adding an item in the backend using the `Cart` tab.

The existing method `#create_proposed_shipments` cannot be used, as the order
is completed, so this new method creates only the missing shipments for the
newly added line item.